+++ src/main/java/kitchenpos/menugroups/infra/JpaMenuGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.menus.infra;
+package kitchenpos.menugroups.infra;
 
 import java.util.UUID;
-import kitchenpos.menus.domain.MenuGroupRepository;
-import kitchenpos.menus.domain.tobe.menugroup.MenuGroup;
+import kitchenpos.menugroups.domain.MenuGroupRepository;
+import kitchenpos.menugroups.domain.tobe.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuGroupRepository extends MenuGroupRepository,
