+++ src/main/java/kitchenpos/menus/application/MenuProductsService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import kitchenpos.menus.domain.tobe.menu.MenuPrice;
-import kitchenpos.menus.domain.tobe.menu.MenuProduct;
-import kitchenpos.menus.domain.tobe.menu.MenuProducts;
+import kitchenpos.menus.domain.tobe.MenuPrice;
+import kitchenpos.menus.domain.tobe.MenuProduct;
+import kitchenpos.menus.domain.tobe.MenuProducts;
 import kitchenpos.menus.ui.dto.MenuProductCreateRequests;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.tobe.Product;
