+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -3,12 +3,12 @@ package kitchenpos.menus.application;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
-import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menugroups.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.menus.domain.tobe.menu.Menu;
-import kitchenpos.menus.domain.tobe.menu.MenuPrice;
-import kitchenpos.menus.domain.tobe.menu.MenuProducts;
-import kitchenpos.menus.domain.tobe.menugroup.MenuGroup;
+import kitchenpos.menus.domain.tobe.Menu;
+import kitchenpos.menus.domain.tobe.MenuPrice;
+import kitchenpos.menus.domain.tobe.MenuProducts;
+import kitchenpos.menugroups.domain.tobe.MenuGroup;
 import kitchenpos.menus.ui.dto.MenuCreateRequest;
 import kitchenpos.products.domain.ProfanityValidator;
 import org.springframework.stereotype.Service;
