+++ src/main/java/kitchenpos/menus/domain/tobe/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain.tobe.menu;
+package kitchenpos.menus.domain.tobe;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Embeddable;
