+++ src/main/java/kitchenpos/menus/infra/JpaMenuRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.menus.infra;
 import java.util.List;
 import java.util.UUID;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.menus.domain.tobe.menu.Menu;
+import kitchenpos.menus.domain.tobe.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
