+++ src/main/java/kitchenpos/menus/ui/MenuRestController.java
@@ -4,8 +4,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.UUID;
 import kitchenpos.menus.application.MenuService;
-import kitchenpos.menus.domain.tobe.menu.Menu;
-import kitchenpos.menus.domain.tobe.menu.MenuPrice;
+import kitchenpos.menus.domain.tobe.Menu;
+import kitchenpos.menus.domain.tobe.MenuPrice;
 import kitchenpos.menus.ui.dto.MenuCreateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
