+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -2,14 +2,14 @@ package kitchenpos.fixture;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.menus.domain.tobe.menu.DisplayedMenu;
-import kitchenpos.menus.domain.tobe.menu.Menu;
-import kitchenpos.menus.domain.tobe.menu.MenuName;
-import kitchenpos.menus.domain.tobe.menu.MenuPrice;
-import kitchenpos.menus.domain.tobe.menu.MenuProduct;
-import kitchenpos.menus.domain.tobe.menu.MenuProducts;
-import kitchenpos.menus.domain.tobe.menu.MenuQuantity;
-import kitchenpos.menus.domain.tobe.menugroup.MenuGroup;
+import kitchenpos.menus.domain.tobe.DisplayedMenu;
+import kitchenpos.menus.domain.tobe.Menu;
+import kitchenpos.menus.domain.tobe.MenuName;
+import kitchenpos.menus.domain.tobe.MenuPrice;
+import kitchenpos.menus.domain.tobe.MenuProduct;
+import kitchenpos.menus.domain.tobe.MenuProducts;
+import kitchenpos.menus.domain.tobe.MenuQuantity;
+import kitchenpos.menugroups.domain.tobe.MenuGroup;
 import kitchenpos.menus.ui.dto.MenuCreateRequest;
 import kitchenpos.menus.ui.dto.MenuProductCreateRequests;
 import kitchenpos.products.domain.tobe.Product;
