+++ src/test/java/kitchenpos/menus/application/MenuProductsServiceTest.java
@@ -8,8 +8,8 @@ import java.util.List;
 import kitchenpos.fake.InMemoryProductRepository;
 import kitchenpos.fixture.MenuProductFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.menus.domain.tobe.menu.MenuPrice;
-import kitchenpos.menus.domain.tobe.menu.MenuProducts;
+import kitchenpos.menus.domain.tobe.MenuPrice;
+import kitchenpos.menus.domain.tobe.MenuProducts;
 import kitchenpos.menus.ui.dto.MenuProductCreateRequests;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.tobe.Product;
