+++ src/test/java/kitchenpos/orders/eatin/application/OrderTableServiceTest.java
@@ -17,10 +17,10 @@ import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.OrderFixture;
 import kitchenpos.fixture.OrderTableFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menugroups.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.menus.domain.tobe.menu.Menu;
-import kitchenpos.menus.domain.tobe.menugroup.MenuGroup;
+import kitchenpos.menus.domain.tobe.Menu;
+import kitchenpos.menugroups.domain.tobe.MenuGroup;
 import kitchenpos.orders.common.domain.Order;
 import kitchenpos.orders.common.domain.OrderRepository;
 import kitchenpos.orders.common.domain.OrderStatus;
